@@ -1,5 +1,5 @@
-const { createProxyMiddleware } = require('http-proxy-middleware');
-module.exports = function (app) {
+import { createProxyMiddleware } from 'http-proxy-middleware';
+export default function(app) {
     app.use(
         createProxyMiddleware('/backend', {
             target: 'http://localhost:5000',
